Make valid_bitmap public
Move valid_bitmap to a separate file and make it public. The valid
bitmap would be usefull also for tracking guest validated pages.
