+++ kernel/src/utils/mod.rs
@@ -10,6 +10,7 @@ pub mod immut_after_init;
 pub mod memory_region;
 pub mod scoped;
 pub mod util;
+pub mod valid_bitmap;
 
 pub use memory_region::MemoryRegion;
 pub use scoped::{ScopedMut, ScopedRef};
